deleted:    click_repl/__init__.py 	deleted:    click_repl/exceptions.py 	deleted:    click_repl/py.typed 	modified:   pyproject.toml 	modified:   setup.cfg 	modified:   tests/conftest.py 	deleted:    tests/test_argument.py 	modified:   tests/test_complete.py 	deleted:    tests/test_dev.py 	new file:   tests/test_dev/test_get_internal_cmds.py 	new file:   tests/test_dev/test_internal_cmds.py 	new file:   tests/test_dev/test_register_internal_cmds.py 	new file:   tests/test_dev/test_sys_cmds.py 	new file:   tests/test_repl.py 	deleted:    tests/test_repl/test_repl.py
