--- click_repl/exceptions.py
@@ -1,6 +1,0 @@
-class InternalCommandException(Exception):
-    pass
-
-
-class ExitReplException(InternalCommandException):
-    pass
