+++ pyproject.toml
@@ -4,8 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [tool.pytest.ini_options]
 addopts = [
-    "--cov=click_repl",
-    "--cov-report=html"
+    "--cov=click_repl"
 ]
 
 testpaths = [
