+++ setup.cfg
@@ -45,8 +45,8 @@ testing =
 [options.package_data]
 click_repl = py.typed
 
-[flake8]
-ignore = E203, E266, W503, E402, E731
-max-line-length = 90
-max-complexity = 18
-select = B,C,E,F,W,T4,B9
+; [flake8]
+; ignore = E203, E266, W503, E402, E731
+; max-line-length = 90
+; max-complexity = 18
+; select = B,C,E,F,W,T4,B9
