+++ tests/conftest.py
@@ -2,11 +2,11 @@ import click_repl
 import pytest
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture
 def send_stdin_input(monkeypatch):
-    iterator = iter((":exit", ":quit", ":q", ":help", ":h", ":?"))
+    iterator = iter((":help", ":h", ":?", ":exit"))
 
-    def fake_stdin(*args):
-        return next(iterator)
+    def fake_stdin_provider(*args):
+        return lambda: next(iterator)
 
-    click_repl.utils._get_command_func = fake_stdin
+    monkeypatch.setattr(click_repl.utils, '_get_command_func', fake_stdin_provider)
